+++ src/test/resources/eclim_unit_test_java/src/org/eclim/test/include/TestImportMissing.java
@@ -1,0 +1,8 @@
+package org.eclim.test.include;
+
+public class TestImportMissing
+{
+  private List list;
+  private ArrayList alist;
+  private FooBar bar;
+}