+++ src/test/resources/eclim_unit_test_java/src/org/eclim/test/include/TestImportMissingVUnit.java
@@ -1,0 +1,7 @@
+package org.eclim.test.include;
+
+public class TestImportMissingVUnit
+{
+  private ArrayList alist;
+  private HashMap amap;
+}